packaging: setup: ssl2jkstrust: do not try to rename if no source
Related-To: https://bugzilla.redhat.com/show_bug.cgi?id=1058016

+++ packaging/ssl2jkstrust.py
@@ -111,7 +111,8 @@ def main():
         p.communicate(input=c)
         if p.returncode != 0:
             raise RuntimeError('keytool failed')
-    os.rename(tmp, options.keystore)
+    if os.path.exists(tmp):
+        os.rename(tmp, options.keystore)
 
 main()
